style(drawer): pref transfrom animation
