+++ packages/drawer/drawer-wrapper.tsx
@@ -1,7 +1,5 @@
-import { CustomCSSProperties } from '../utils/base'
 import { computed, defineComponent, PropType } from 'vue'
 import { Placement } from './props'
-import { getDrawerTransfrom } from './style'
 
 const DrawerWrapper = defineComponent({
   props: {
@@ -18,18 +16,9 @@ const DrawerWrapper = defineComponent({
       return round && 'round'
     })
 
-    const setDrawerStyle = computed(() => {
-      const { placement } = props
-      const styles: CustomCSSProperties = {
-        '--drawer-transfrom': getDrawerTransfrom(placement),
-      }
-      return styles
-    })
-
     return () => (
       <div
         class={`fect-drawer__wrapper fect-drawer__wrapper--${props.placement} ${setRounder.value}`}
-        style={setDrawerStyle.value}
         role="dialog"
       >
         {slots.default?.()}
